client-admin/requirements.txt
pipecat-ai[silero,websocket,google,local-smart-turn-v3]==0.0.92
fastapi
uvicorn
elevenlabs
websockets
python-dotenv
python-multipart
aiohttp
sqlalchemy
aiomysql
asyncpg
aiosmtplib
google-generativeai
